Rewrote collision system yet again. I also corrected the physics system to the following units: -1 Block = 1 Meter -Acceleration of gravity =  +18 m/s^2 y is now a float. itemMob Graphics are no longer allocated per itemMob to prevent reloading graphics. baseMob pointers are now set to NULL after deletion to prevent re-deletion of non-NULL pointers. Mobs are now controlled horizontally via velocity.
Notes:
-Since the unit change, horizontal velocities are too slow.
-Too many itemMobs still causes a dramatic decrease in FPS.

+++ source/mobs/baseMob.h
@@ -7,7 +7,7 @@ bool canBaseMobSpawnHere(WorldObject* world, int x, int y);
 class baseMob {
 public:
 	float x;
-	int y;
+	float y;
 	float vy;
 	float vx;
 	int sx;
